@@ -7,23 +7,23 @@
     "babel-plugin-transform-async-to-generator": "^6.24.1",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.5.1",
+    "babel-preset-env": "^1.5.2",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-flow": "^6.23.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
     "babel-register": "^6.24.1",
     "babel-runtime": "^6.23.0",
-    "create-react-class": "^15.5.2",
-    "eslint": "^3.19.0",
-    "eslint-config-google": "^0.7.1",
+    "create-react-class": "^15.6.0",
+    "eslint": "^4.0.0",
+    "eslint-config-google": "^0.8.0",
     "eslint-config-prettier": "^2.1.1",
     "eslint-plugin-flow-vars": "^0.5.0",
     "eslint-plugin-flowtype": "^2.32.1",
     "eslint-plugin-import": "^2.3.0",
     "eslint-plugin-jsx-a11y": "^5.0.3",
-    "eslint-plugin-prettier": "^2.1.1",
-    "eslint-plugin-react": "^7.0.0",
+    "eslint-plugin-prettier": "^2.1.2",
+    "eslint-plugin-react": "^7.1.0",
     "flow-bin": "^0.42.0",
     "glob": "^7.1.2",
     "iflow-debug": "^1.0.15",
