@@ -1,7 +1,7 @@
 {
   "name": "gatsby",
   "description": "React.js Static Site Generator",
-  "version": "1.0.0-alpha.22",
+  "version": "1.0.0-alpha.23",
   "author": "Kyle Mathews <mathews.kyle@gmail.com>",
   "bin": {
     "gatsby": "./dist/gatsby-cli.js"
@@ -23,18 +23,18 @@
     "babel-preset-stage-0": "^6.24.1",
     "babel-runtime": "^6.23.0",
     "babel-traverse": "^6.24.1",
-    "babylon": "^6.17.0",
+    "babylon": "^6.17.3",
     "bluebird": "^3.5.0",
     "boom": "^2.7.2",
     "bundle-loader": "^0.5.4",
     "chalk": "^1.1.3",
-    "chokidar": "^1.6.1",
+    "chokidar": "^1.7.0",
     "chunk-manifest-webpack-plugin": "0.1.0",
     "commander": "^2.9.0",
     "copyfiles": "^1.2.0",
     "css-loader": "^0.26.1",
     "debug": "^2.6.0",
-    "detect-port": "^1.1.1",
+    "detect-port": "^1.2.1",
     "eventemitter2": "^4.1.0",
     "express": "^4.14.0",
     "express-graphql": "^0.6.1",
@@ -48,7 +48,7 @@
     "graphql-skip-limit": "^1.0.0-alpha.22",
     "gray-matter": "^2.1.0",
     "hapi": "^8.5.1",
-    "history": "^4.6.1",
+    "history": "^4.6.2",
     "hosted-git-info": "^2.4.2",
     "invariant": "^2.2.2",
     "is-relative": "^0.2.1",
@@ -56,22 +56,24 @@
     "joi": "^9.1.1",
     "json-loader": "^0.5.2",
     "json5": "^0.5.0",
-    "level": "^1.6.0",
     "loader-utils": "^0.2.16",
     "lodash": "^4.17.4",
-    "markdown-it": "^7.0.1",
+    "lodash-id": "^0.14.0",
+    "lowdb": "^0.16.2",
     "md5-file": "^3.1.1",
     "mime": "^1.3.4",
     "mime-types": "^2.1.15",
+    "mitt": "^1.1.2",
     "mkdirp": "^0.5.1",
     "moment": "^2.16.0",
     "negotiator": "^0.6.1",
     "node-cjsx": "^2.0.0",
     "node-libs-browser": "^2.0.0",
+    "normalize-path": "^2.1.1",
     "null-loader": "^0.1.1",
     "object-assign": "^4.1.0",
     "observable": "^2.1.4",
-    "opn": "^5.0.0",
+    "opn": "^5.1.0",
     "parse-filepath": "^1.0.1",
     "path-exists": "^3.0.0",
     "postcss-browser-reporter": "^0.5.0",
@@ -80,8 +82,8 @@
     "postcss-loader": "^0.13.0",
     "postcss-reporter": "^1.4.1",
     "raw-loader": "^0.5.1",
-    "react": "^15.4.0",
-    "react-dom": "^15.4.0",
+    "react": "^15.6.0",
+    "react-dom": "^15.6.0",
     "react-hot-loader": "^3.0.0-beta.6",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
