@@ -1,9 +1,16 @@
-import React from "react"
+import React, { Component } from "react"
+import * as PropTypes from "prop-types"
 import { TypographyStyle } from "react-typography"
 
 import typography from "./utils/typography"
 
-module.exports = React.createClass({
+const propTypes = {
+  headComponents: PropTypes.node.isRequired,
+  body: PropTypes.node.isRequired,
+  postBodyComponents: PropTypes.node.isRequired,
+}
+
+class Html extends Component {
   render() {
     return (
       <html op="news" lang="en">
@@ -33,5 +40,9 @@ module.exports = React.createClass({
         </body>
       </html>
     )
-  },
-})
+  }
+}
+
+Html.propTypes = propTypes
+
+module.exports = Html
