@@ -1,4 +1,5 @@
-import React from "react"
+import React, { Component } from "react"
+import * as PropTypes from "prop-types"
 import { TypographyStyle } from "react-typography"
 import typography from "./utils/typography"
 
@@ -11,7 +12,13 @@ if (process.env.NODE_ENV === `production`) {
   }
 }
 
-module.exports = React.createClass({
+const propTypes = {
+  headComponents: PropTypes.node.isRequired,
+  body: PropTypes.node.isRequired,
+  postBodyComponents: PropTypes.node.isRequired,
+}
+
+class Html extends Component {
   render() {
     let css
     if (process.env.NODE_ENV === `production`) {
@@ -49,5 +56,9 @@ module.exports = React.createClass({
         </body>
       </html>
     )
-  },
-})
+  }
+}
+
+Html.propTypes = propTypes
+
+module.exports = Html
