@@ -1,5 +1,5 @@
-declare var __PREFIX_LINKS__: bool;
-declare var __LINK_PREFIX__: string;
+declare var __PREFIX_PATHS__: bool;
+declare var __PATH_PREFIX__: string;
 declare module 'gatsby-helpers' {
   declare function prefixLink(link: string): string;
 }
