@@ -1,25 +1,27 @@
-const path = require('path')
+const path = require(`path`)
 
 exports.onCreateNode = ({ node, boundActionCreators, getNode }) => {
   const { createNodeField } = boundActionCreators
 
-  if (node.internal.type === 'MarkdownRemark') {
+  if (node.internal.type === `MarkdownRemark`) {
     const fileNode = getNode(node.parent)
-    let nodeSlug 
-    nodeSlug = ensureSlashes(path.basename(fileNode.relativePath, path.extname(fileNode.relativePath)))
+    let nodeSlug
+    nodeSlug = ensureSlashes(
+      path.basename(fileNode.relativePath, path.extname(fileNode.relativePath))
+    )
     if (nodeSlug) {
-      createNodeField({ node, fieldName: 'slug', fieldValue: nodeSlug })
+      createNodeField({ node, fieldName: `slug`, fieldValue: nodeSlug })
     }
   }
 }
 
 function ensureSlashes(slug) {
-  if (slug.charAt(0) !== '/') {
-    slug = '/' + slug
+  if (slug.charAt(0) !== `/`) {
+    slug = `/` + slug
   }
 
-  if (slug.charAt(slug.length -1) !== '/') {
-    slug = slug + '/'
+  if (slug.charAt(slug.length - 1) !== `/`) {
+    slug = slug + `/`
   }
 
   return slug
