@@ -1,6 +1,6 @@
-import React from 'react'
+import React from "react"
 
-const IndexRoute = () => (
+const IndexRoute = () =>
   <div>
     <p>
       Welcome to the GatsbyJS RSS Demo.
@@ -8,7 +8,5 @@ const IndexRoute = () => (
       to see the generated RSS Feed.
     </p>
   </div>
-)
-
 
 export default IndexRoute
