@@ -1,14 +1,13 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import Link from 'gatsby-link'
-import { connect } from 'react-redux'
+import React from "react"
+import PropTypes from "prop-types"
+import Link from "gatsby-link"
+import { connect } from "react-redux"
 
-const Counter = ({ count, increment }) => (
+const Counter = ({ count, increment }) =>
   <div>
     <p>Count: {count}</p>
     <button onClick={increment}>Increment</button>
   </div>
-)
 
 Counter.propTypes = {
   count: PropTypes.number.isRequired,
@@ -19,7 +18,7 @@ const mapStateToProps = ({ count }) => {
   return { count }
 }
 
-const mapDispatchToProps = (dispatch) => {
+const mapDispatchToProps = dispatch => {
   return { increment: () => dispatch({ type: `INCREMENT` }) }
 }
 
@@ -34,7 +33,7 @@ class DefaultLayout extends React.Component {
             Redux example
           </h3>
         </Link>
-        <ConnectedCounter/>
+        <ConnectedCounter />
         <ul>
           <li><Link to="/a/">a</Link></li>
           <li><Link to="/b/">b</Link></li>
