@@ -1,12 +1,12 @@
-import { createStore as reduxCreateStore } from 'redux'
+import { createStore as reduxCreateStore } from "redux"
 
 const reducer = (state, action) => {
-    if (action.type === `INCREMENT`) {
-        return Object.assign({}, state, {
-            count: state.count + 1,
-        })
-    }
-    return state
+  if (action.type === `INCREMENT`) {
+    return Object.assign({}, state, {
+      count: state.count + 1,
+    })
+  }
+  return state
 }
 
 const initialState = { count: 0 }
