@@ -1,13 +1,11 @@
-import React from 'react'
+import React from "react"
 
-const IndexRoute = () => (
+const IndexRoute = () =>
   <div>
     <p>
       Welcome to the GatsbyJS Sitemap Demo.
       Visit <a href="/sitemap.xml">to see the generated sitemap.</a>
     </p>
   </div>
-)
-
 
 export default IndexRoute
