@@ -2,39 +2,37 @@ import React from "react"
 import { Link } from "react-router-dom"
 import PropTypes from "prop-types"
 
-if (typeof window !== `undefined`) {
-  require(`ric`)
-}
-
-let linkPrefix = ``
-if (__PREFIX_LINKS__) {
-  linkPrefix = __LINK_PREFIX__
+let pathPrefix = ``
+if (__PREFIX_PATHS__) {
+  pathPrefix = __PATH_PREFIX__
 }
 
 class GatsbyLink extends React.Component {
+  constructor(props) {
+    super()
+    this.state = {
+      to: pathPrefix + props.to,
+    }
+  }
   propTypes: {
     to: PropTypes.string.isRequired,
     onClick: PropTypes.func,
   }
-  componentDidMount() {
-    // Only enable prefetching of Link resources in production and for browsers
-    // that don't support service workers *cough* Safari/IE *cough*.
-    //
-    // TODO also add check if user is using SW, e.g. window.caches as if
-    // not we should preload here too.
-    if (
-      process.env.NODE_ENV === `production` &&
-      (!(`serviceWorker` in window.navigator) ||
-        window.location.protocol !== `https:`)
-    ) {
-      requestUserIdle(() => {
-        ___loadScriptsForPath(this.props.to)
+
+  componentWillReceiveProps(nextProps) {
+    if (this.props.to !== nextProps.to) {
+      this.setState({
+        to: pathPrefix + props.to,
       })
+      ___loader.enqueue(this.state.to)
     }
   }
 
+  componentDidMount() {
+    ___loader.enqueue(this.state.to)
+  }
+
   render() {
-    const to = linkPrefix + this.props.to
     const { onClick, ...rest } = this.props
     return (
       <Link
@@ -42,9 +40,12 @@ class GatsbyLink extends React.Component {
           onClick && onClick(e)
           // Is this link pointing to a hash on the same page? If so,
           // just scroll there.
-          const pathname = this.props.to.split(`#`).slice(0, -1).join(``)
+          let pathname = this.state.to
+          if (pathname.split(`#`).length > 1) {
+            pathname = pathname.split(`#`).slice(0, -1).join(``)
+          }
           if (pathname === window.location.pathname) {
-            const hashFragment = this.props.to.split(`#`).slice(1).join(`#`)
+            const hashFragment = this.state.to.split(`#`).slice(1).join(`#`)
             const element = document.getElementById(hashFragment)
             if (element !== null) {
               element.scrollIntoView()
@@ -56,11 +57,11 @@ class GatsbyLink extends React.Component {
           // loaded before continuing.
           if (process.env.NODE_ENV === `production`) {
             e.preventDefault()
-            window.___navigateTo(this.props.to)
+            window.___navigateTo(this.state.to)
           }
         }}
         {...rest}
-        to={to}
+        to={this.state.to}
       />
     )
   }
