@@ -1,20 +1,20 @@
-import path from 'path'
-import { defaultOptions, runQuery, writeFile } from './internals'
+import path from "path"
+import RSS from "RSS"
+import { defaultOptions, runQuery, writeFile } from "./internals"
 
-const publicPath = './public'
+const publicPath = `./public`
 
 // A default function to transform query data into feed entries.
-const serialize = ({ site, allMarkdownRemark }) => (
-  allMarkdownRemark.edges.map(edge => ({
-    ...edge.node.frontmatter,
-    description: edge.node.excerpt,
-    url: site.siteMetadata.site_url + edge.node.fields.slug,
-    guid: site.siteMetadata.site_url + edge.node.fields.slug,
-    custom_elements: [
-      { 'content:encoded': edge.node.html }
-    ]
-  }))
-)
+const serialize = ({ site, allMarkdownRemark }) =>
+  allMarkdownRemark.edges.map(edge => {
+    return {
+      ...edge.node.frontmatter,
+      description: edge.node.excerpt,
+      url: site.siteMetadata.site_url + edge.node.fields.slug,
+      guid: site.siteMetadata.site_url + edge.node.fields.slug,
+      custom_elements: [{ "content:encoded": edge.node.html }],
+    }
+  })
 
 exports.onPostBuild = async ({ graphql }, pluginOptions) => {
   delete pluginOptions.plugins
@@ -38,7 +38,7 @@ exports.onPostBuild = async ({ graphql }, pluginOptions) => {
 
     const output = path.join(publicPath, f.output)
     const ctx = { ...globals, ...locals }
-    const feed = setup({ ...rest, ...ctx })
+    const feed = new RSS(setup({ ...rest, ...ctx }))
     const items = f.serialize ? f.serialize(ctx) : serialize(ctx)
 
     items.forEach(i => feed.item(i))
