@@ -1,13 +1,12 @@
-import fs from 'fs'
-import RSS from 'rss'
-import pify from 'pify'
+import fs from "fs"
+import pify from "pify"
 
 export const writeFile = pify(fs.writeFile)
 
-export const runQuery = (handler, query) => handler(query)
-  .then(r => {
+export const runQuery = (handler, query) =>
+  handler(query).then(r => {
     if (r.errors) {
-      throw new Error(r.errors.join(', '))
+      throw new Error(r.errors.join(`, `))
     }
 
     return r.data
@@ -31,11 +30,17 @@ export const defaultOptions = {
   `,
 
   // Setup a few RSS object, merging on various feed-speciupfic options.
-  setup: ({ site: { siteMetadata }, ...rest }) => new RSS({ ...siteMetadata, ...rest }),
+  setup: ({ site: { siteMetadata }, ...rest }) => {
+    return {
+      ...siteMetadata,
+      ...rest,
+    }
+  },
 
   // Create a default RSS feed. Others may be added by using the format below.
-  feeds: [{
-    query: `
+  feeds: [
+    {
+      query: `
       {
         allMarkdownRemark(
           limit: 1000,
@@ -64,7 +69,8 @@ export const defaultOptions = {
       }
     `,
 
-    // Where we will save the feed generated by this query.
-    output: `rss.xml`,
-  }]
+      // Where we will save the feed generated by this query.
+      output: `rss.xml`,
+    },
+  ],
 }
