@@ -18,7 +18,9 @@ exports.onPostBuild = (args, pluginOptions) => {
 
   const options = {
     staticFileGlobs: [
-      `${rootDir}/**/*.{js,woff2}`,
+      `${rootDir}/**/*.{woff2}`,
+      `${rootDir}/commons-*js`,
+      `${rootDir}/app-*js`,
       `${rootDir}/index.html`,
       `${rootDir}/manifest.json`,
       `${rootDir}/offline-plugin-app-shell-fallback/index.html`,
@@ -40,7 +42,7 @@ exports.onPostBuild = (args, pluginOptions) => {
     runtimeCaching: [
       {
         // Add runtime caching of images.
-        urlPattern: /\.(?:png|jpg|jpeg|webp|svg|gif|tiff)$/,
+        urlPattern: /\.(?:png|jpg|jpeg|webp|svg|gif|tiff|js|woff|woff2)$/,
         handler: `fastest`,
       },
     ],
