@@ -23,10 +23,10 @@ module.exports = ({ files, markdownNode, markdownAST, getNode }) => {
         const newPath = path.join(
           process.cwd(),
           `public`,
-          `${linkNode.contentDigest}.${linkNode.extension}`
+          `${linkNode.internal.contentDigest}.${linkNode.extension}`
         )
         const relativePath = path.join(
-          `/${linkNode.contentDigest}.${linkNode.extension}`
+          `/${linkNode.internal.contentDigest}.${linkNode.extension}`
         )
         link.url = `${relativePath}`
         if (!fsExtra.existsSync(newPath)) {
@@ -44,9 +44,21 @@ module.exports = ({ files, markdownNode, markdownAST, getNode }) => {
     visitor(link)
   })
 
-  // Also copy gifs since Sharp can't process them.
+  // Also copy gifs since Sharp can't process them as well as svgs since we
+  // exclude them from the image processing pipeline in
+  // gatsby-remark-responsive-image
   visit(markdownAST, `image`, image => {
-    if (image.url.slice(-3) === `gif`) {
+    const imagePath = path.join(getNode(markdownNode.parent).dir, image.url)
+    const imageNode = _.find(files, file => {
+      if (file && file.absolutePath) {
+        return file.absolutePath === imagePath
+      }
+      return false
+    })
+    if (
+      imageNode &&
+      (imageNode.extension === `gif` || imageNode.extension === `svg`)
+    ) {
       visitor(image)
     }
   })
