@@ -11,14 +11,10 @@ const Promise = require(`bluebird`)
 // 3. Filter out any responsive image sizes that are greater than the image's width
 // 4. Create the responsive images.
 // 5. Set the html w/ aspect ratio helper.
-module.exports = ({
-  files,
-  markdownNode,
-  markdownAST,
-  pluginOptions,
-  linkPrefix,
-  getNode,
-}) => {
+module.exports = (
+  { files, markdownNode, markdownAST, pathPrefix, getNode },
+  pluginOptions
+) => {
   const defaults = {
     maxWidth: 650,
     wrapperStyle: ``,
@@ -31,8 +27,15 @@ module.exports = ({
     imageNodes.map(
       node =>
         new Promise((resolve, reject) => {
-          // Ignore gifs as we can't process them.
-          if (isRelativeUrl(node.url) && node.url.slice(-3) !== `gif`) {
+          const fileType = node.url.slice(-3)
+
+          // Ignore gifs as we can't process them,
+          // svgs as they are already responsive by definition
+          if (
+            isRelativeUrl(node.url) &&
+            fileType !== `gif` &&
+            fileType !== `svg`
+          ) {
             const imagePath = path.posix.join(
               getNode(markdownNode.parent).dir,
               node.url
@@ -73,13 +76,13 @@ module.exports = ({
             style="display: block"
             target="_blank"
           >
-            <div
+            <span
               class="gatsby-resp-image-wrapper"
-              style="position: relative; z-index: -1; ${options.wrapperStyle}"
+              style="position: relative; z-index: -1; display: block; ${options.wrapperStyle}"
             >
-              <div
+              <span
                 class="gatsby-resp-image-background-image"
-                style="padding-bottom: ${ratio};position: relative; width: 100%; bottom: 0; left: 0; background-image: url('${responsiveSizesResult.src}'); background-size: cover;"
+                style="padding-bottom: ${ratio};position: relative; width: 100%; bottom: 0; left: 0; background-image: url('${responsiveSizesResult.base64}'); background-size: cover; display: block;"
               >
                 <img
                   class="gatsby-resp-image-image"
@@ -90,22 +93,14 @@ module.exports = ({
                   srcset="${srcSet}"
                   sizes="${responsiveSizesResult.sizes}"
                 />
-              </div>
-            </div>
+              </span>
+            </span>
           </a>
           `
-              // const rawHTML = `
-              // <div style="width: ${responsiveSizesResult.width}px; height: ${responsiveSizesResult.height}px;  padding-bottom: ${responsiveSizesResult.aspectRatio * 100}%;" class="image-loader">
-              // `
 
-              node.data = {
-                hChildren: [{ type: `raw`, value: rawHTML }],
-              }
-              // Set type to unknown so mdast-util-to-hast will treat this node as a
-              // div not an image — it gets quite confused otherwise and tries to put
-              // the raw html above as a child of the image which browsers
-              // justifiably squawk at.
-              node.type = `unknown`
+              // Replace the image node with an inline HTML node.
+              node.type = `html`
+              node.value = rawHTML
               return resolve()
             })
           } else {
