@@ -49,8 +49,8 @@ program
   .command(`build`)
   .description(`Build a Gatsby project.`)
   .option(
-    `--prefix-links`,
-    `Build site with links prefixed (set prefix in your config).`
+    `--prefix-paths`,
+    `Build site with link paths prefixed (set prefix in your config).`
   )
   .action(command => {
     // Set NODE_ENV to 'production'
@@ -62,7 +62,7 @@ program
       directory,
     }
     build(p).then(() => {
-      console.log(`Done`)
+      console.log(`Done building in`, process.uptime(), `seconds`)
       process.exit()
     })
   })
