@@ -15,13 +15,15 @@ const { store } = require(`../../redux/`)
 const { boundActionCreators } = require(`../../redux/actions`)
 const queryCompiler = require(`./query-compiler`).default
 const queryRunner = require(`./query-runner`)
+const invariant = require(`invariant`)
+const normalize = require(`normalize-path`)
 
 exports.extractQueries = () => {
   const pages = store.getState().pages
   const components = _.uniq(pages.map(p => p.component))
   return queryCompiler().then(queries => {
     components.forEach(component => {
-      const query = queries.get(component)
+      const query = queries.get(normalize(component))
 
       boundActionCreators.replacePageComponentQuery({
         query: query && query.text,
@@ -57,6 +59,11 @@ exports.watch = rootDir => {
     queryCompiler().then(queries => {
       const pages = store.getState().pageComponents
       queries.forEach(({ text }, path) => {
+        invariant(
+          pages[path],
+          `Path ${path} not found in the store pages: ${JSON.stringify(pages)}`
+        )
+
         if (text !== pages[path].query) {
           boundActionCreators.replacePageComponentQuery({
             query: text,
