@@ -5,21 +5,24 @@ const { GraphQLSchema, GraphQLObjectType } = require(`graphql`)
 const buildNodeTypes = require(`./build-node-types`)
 const buildNodeConnections = require(`./build-node-connections`)
 const { store } = require(`../redux`)
+const invariant = require(`invariant`)
 
 module.exports = async () => {
+  console.time(`building schema`)
+
   const typesGQL = await buildNodeTypes()
   const connections = buildNodeConnections(_.values(typesGQL))
+
+  // Pull off just the graphql node from each type object.
+  const nodes = _.mapValues(typesGQL, `node`)
+
+  invariant(!_.isEmpty(nodes), `There are no available GQL nodes`)
+  invariant(!_.isEmpty(connections), `There are no available GQL connections`)
 
   const schema = new GraphQLSchema({
     query: new GraphQLObjectType({
       name: `RootQueryType`,
-      fields: () => {
-        return {
-          // Pull off just the graphql node from each type object.
-          ..._.mapValues(typesGQL, `node`),
-          ...connections,
-        }
-      },
+      fields: { ...nodes, ...connections },
     }),
   })
 
@@ -27,6 +30,4 @@ module.exports = async () => {
     type: `SET_SCHEMA`,
     payload: schema,
   })
-
-  return
 }
